+++ src/probabilities_estimators/Shrinkage.jl
@@ -49,7 +49,7 @@ For [`allprobabilities`](@ref), `m = total_outcomes(o, x)`, where `o` is the
 ```julia
 using ComplexityMeasures
 x = cumsum(randn(100))
-ps_shrink = probabilities(Shrinkage(SymbolicPermutation(m = 3)), x)
+ps_shrink = probabilities(Shrinkage(OrdinalPatterns(m = 3)), x)
 ```
 
 See also: [`RelativeAmount`](@ref), [`Bayes`](@ref).