+++ test/outcome_spaces/implementations/spatial/spatial_ordinal_patterns.jl
@@ -7,7 +7,7 @@ x = [1 2 1; 8 3 4; 6 7 5]
 # (you get 4 symbols in a 3x3 matrix. For this matrix, the upper left
 # and bottom right are the same symbol. So three probabilities in the end).
 stencil = CartesianIndex.([(0,0), (1,0), (0,1), (1,1)])
-est = SpatialSymbolicPermutation(stencil, x; periodic = false)
+est = SpatialOrdinalPatterns(stencil, x; periodic = false)
 
 # Generic tests
 ps = probabilities(est, x)
@@ -19,13 +19,13 @@ h = information(Renyi(base = 2), est, x)
 # In fact, doesn't matter how we order the stencil,
 # the symbols will always be equal in top-left and bottom-right
 stencil = CartesianIndex.([(0,0), (1,0), (1,1), (0,1)])
-est = SpatialSymbolicPermutation(stencil, x; periodic = false)
+est = SpatialOrdinalPatterns(stencil, x; periodic = false)
 @test information(Renyi(base = 2), est, x) == 1.5
 
 # But for sanity, let's ensure we get a different number
 # for a different stencil
 stencil = CartesianIndex.([(0,0), (1,0), (2,0)])
-est = SpatialSymbolicPermutation(stencil, x; periodic = false)
+est = SpatialOrdinalPatterns(stencil, x; periodic = false)
 ps = sort(probabilities(est, x))
 @test ps[1] == 1/3
 @test ps[2] == 2/3
@@ -33,12 +33,12 @@ ps = sort(probabilities(est, x))
 # if we specify extent and lag instead of stencil
 extent = (2, 2)
 lag = (1, 1)
-est = SpatialSymbolicPermutation((extent, lag), x; periodic = false)
+est = SpatialOrdinalPatterns((extent, lag), x; periodic = false)
 @test information(Renyi(base = 2), est, x) == 1.5
 
 # and let's also test the matrix-way of specifying the stencil
 stencil = [1 1; 1 1];
-est = SpatialSymbolicPermutation(stencil, x; periodic = false)
+est = SpatialOrdinalPatterns(stencil, x; periodic = false)
 @test information(Renyi(base = 2), est, x) == 1.5
 # when the stencil is square, it is also easy to get an analytical set of outcomes.
 # 1 2 1    count column major order and get vectors, starting in top left corner,
@@ -52,7 +52,7 @@ ps, outs = probabilities_and_outcomes(est, x)
 # Also test that it works for arbitrarily high-dimensional arrays
 stencil = CartesianIndex.([(0,0,0), (0,1,0), (0,0,1), (1,0,0)])
 z = reshape(1:125, (5,5,5));
-est = SpatialSymbolicPermutation(stencil, z; periodic = false)
+est = SpatialOrdinalPatterns(stencil, z; periodic = false)
 # Analytically the total stencils are of length 4*4*4 = 64
 # but all of them given the same probabilities because of the layout
 ps = probabilities(est, z)
@@ -66,16 +66,16 @@ ps = probabilities(est, w)
 stencil = CartesianIndex.([(0,0,0), (0,1,0), (0,0,1),
                             (1,0,0), (0,1,1), (1,0,1),
                             (1,1,0), (1,1,1)])
-est1 = SpatialSymbolicPermutation(stencil, w; periodic = false)
+est1 = SpatialOrdinalPatterns(stencil, w; periodic = false)
 # which would correspond to this
 extent = (2, 2, 2)
 lag = (1, 1, 1)
-est2 = SpatialSymbolicPermutation((extent, lag), w; periodic = false)
+est2 = SpatialOrdinalPatterns((extent, lag), w; periodic = false)
 @test information(Renyi(), est1, w) == information(Renyi(), est2, w)
 
 # and to this stencil written as a matrix
 stencil = [1; 1;; 1; 1;;; 1; 1;; 1; 1]
-est3 = SpatialSymbolicPermutation(stencil, w; periodic = false)
+est3 = SpatialOrdinalPatterns(stencil, w; periodic = false)
 @test information(Renyi(), est1, w, ) == information(Renyi(), est3, w)
 
 ####################
@@ -85,8 +85,8 @@ est3 = SpatialSymbolicPermutation(stencil, w; periodic = false)
 # noise. Test this assumption up to some tolerance.
 x = rand(100, 100)
 stencil = [1 1; 1 1];
-est = SpatialSymbolicPermutation(stencil, x)
+est = SpatialOrdinalPatterns(stencil, x)
 hsp = information_normalized(Renyi(), est, x)
 @test round(hsp, digits = 2) == 1.00
 
-@test outcome_space(est) == outcome_space(SymbolicPermutation(m = 4))
+@test outcome_space(est) == outcome_space(OrdinalPatterns(m = 4))