+++ test/utils/encoding.jl
@@ -227,9 +227,9 @@ end
     #    (4, 2, 1) -> (3, 2, 1)
     #    (2, 1, 0) -> (3, 2, 1),
     # so there are three occurring patterns and m! - 3 = 3*2*1 - 3 = 3 missing patterns
-    @test missing_outcomes(SymbolicPermutation(; m, τ), x) == 3
+    @test missing_outcomes(OrdinalPatterns(; m, τ), x) == 3
 
     m, τ = 2, 1
     y = [1, 2, 1, 2] # only two patterns, none missing
-    @test missing_outcomes(SymbolicPermutation(; m, τ), x) == 0
+    @test missing_outcomes(OrdinalPatterns(; m, τ), x) == 0
 end